GH-87: Remove legacy query widget.
